+++ lib/hcloud/entities/change_protection.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module HCloud
+  class ChangeProtection < Entity
+    attribute :change, :boolean
+
+    alias change? change
+  end
+end